backend/requirements.txt
# Web framework
fastapi==0.111.0
uvicorn[standard]==0.29.0

# Database
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pgvector==0.2.5

# Auth
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
passlib[bcrypt]==1.7.4
python-multipart==0.0.9

# Validation
pydantic==2.7.1
pydantic-settings==2.2.1

# Embeddings
sentence-transformers==3.0.1

# PDF parsing
pdfplumber==0.11.0

# HTTP client (for Ollama)
httpx==0.27.0

# Rate limiting
slowapi==0.1.9

# Config
python-dotenv==1.0.1

# Testing
pytest==8.2.0
pytest-asyncio==0.23.6
httpx==0.27.0
